// src/types/dtos/userDto.ts
export type configUserDto = {
  user_id: string;
  user_passwd: string;
  user_name: string;
  user_div: string;
  dept_code: string;
  email: string;
  tel_no: string;
  team_leader: string;
  mobile_phone: string;
  description: string;
  history_id: string;
  history_type: string;
};

// export type TestCaseListResponseDto = {
//   dto: {
//     TestCaseDto: TestCaseDto[];
//     pageInfoDto: PageInfoDto;
//     pagingResultDto: PagingResultDto;
//   };
// };